requirements.txt
cryptography>=41.0.7
qrcode>=7.3.1
Pillow>=9.5.0
reportlab>=3.6.12
tkcalendar>=1.6.1
opencv-python>=4.8.0.74